requirements.txt
Django==1.8.4
django-leaflet==0.16.0
djangorestframework==3.1.3
django-allauth==0.23.0
requests==2.7.0
django-cors-headers==1.1.0
django-oauth-toolkit==0.9.0
